refactor: single model-catalog hook for the permissions picker (#13494)
- Extracts useModelCategories() from the byte-identical catalog fetch
  effects in authorize.tsx and account-permissions-input.tsx
- AccountPermissionsInput now takes modelCategories as a prop, so the
  authorize screen's grant summary and its checkbox list render from one
  array instead of two independent fetches
- Fixes three dead Scalar anchors: #tag/bring-your-own-pollen → #tag/byop
